+++ addon/models/search-users.js
@@ -1,0 +1,17 @@
+import OsfModel from './user';
+
+/**
+ * @module ember-osf
+ * @submodule models
+ */
+
+/**
+ * Model for OSF APIv2 search/users. This extends the user model.
+ * This model allowed searching users by adding a query param q=<query>
+ * For information on user search:
+ * * https://api.osf.io/v2/search/users/
+ * @class SearchUser
+ */
+
+export default OsfModel.extend({
+});