+++ addon/serializers/search-users.js
@@ -1,0 +1,3 @@
+import OsfSerializer from './osf-serializer';
+
+export default OsfSerializer.extend({});